Extract fetch function
