+++ components/assets/icons/index.js
@@ -1,0 +1,8 @@
+export { default as IconCoffee } from "./Coffee";
+export { default as IconDiscord } from "./Discord";
+export { default as IconGithub } from "./GitHub";
+export { default as IconReddit } from "./Reddit";
+export { default as IconTrello } from "./Trello";
+export { default as IconYoutube } from "./Youtube";
+export { default as IconBluesky } from "./Bluesky";
+export { default as IconTwitter } from "./Xtwitter";
