+++ services/index.js
@@ -1,0 +1,4 @@
+export { fetchCommunityReleasesFromRepos } from "./fetchCommunityAppReleases";
+export { fetchOfficialAppsData } from "./fetchOfficialAppReleases";
+export { fetchTotalDownloadsFromRepo, fetchTotalDownloadsFromRepos } from "./fetchTotalDownloads";
+export { fetchServerReleases } from "./fetchServerReleases";
