+++ utils/titleCase.jsx
@@ -1,0 +1,6 @@
+export const titleCase = (str) => {
+  return str
+    .split('-')
+    .map(word => word.charAt(0).toUpperCase() + word.slice(1).toLowerCase())
+    .join(' ');
+};
